usefull_findings_for_stakeholders
An SQL analytics project for a global music store involved using SQL to analyze data from the store’s database to gain insights into customer behavior, sales trends, and other key performance indicators. The goal of the project is to use data analysis to make informed decisions that can improve the store’s operations and profitability. This involved creating reports and visualizations to present the data in a clear and understandable way, as well as using advanced analytics techniques to uncover hidden patterns and relationships in the data.